+++ src/features/editor/extensions/message/index.ts
@@ -1,6 +1,5 @@
-import { Plugin, PluginKey } from "@tiptap/pm/state";
-import { Decoration, DecorationSet } from "@tiptap/pm/view";
 import { InputRule, mergeAttributes, Node } from "@tiptap/react";
+import { createMessageSymbolDecorationPlugin } from "./message-symbol-decoration-plugin";
 
 export interface MessageOptions {
   type?: "message" | "alert";
@@ -10,6 +9,7 @@ export const Message = Node.create({
   name: "message",
   group: "block",
   content: "messageContent",
+  isolating: true,
 
   addAttributes() {
     return {
@@ -65,32 +65,6 @@ export const Message = Node.create({
   },
 
   addProseMirrorPlugins() {
-    return [
-      new Plugin({
-        key: new PluginKey("messageSymbolDecoration"),
-        props: {
-          decorations: (state) => {
-            // TODO: トランザクションの度に再実行されてるので、パフォーマンス改善の余地あり
-            const { doc } = state;
-            const decorations: Decoration[] = [];
-
-            doc.descendants((node, pos) => {
-              if (node.type.name === this.name) {
-                decorations.push(
-                  Decoration.widget(pos + 1, () => {
-                    const element = document.createElement("span");
-                    element.className = "msg-symbol";
-                    element.textContent = "!";
-                    return element;
-                  }),
-                );
-              }
-            });
-
-            return DecorationSet.create(doc, decorations);
-          },
-        },
-      }),
-    ];
+    return [createMessageSymbolDecorationPlugin(this.name)];
   },
 });
